backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.89.0
groq==0.28.0
requests==2.31.0
email-validator==2.1.0
gunicorn==21.2.0
cryptography==41.0.7